+++ sdk/resourcemanager/datafactory/armdatafactory/constants.go
@@ -10,7 +10,7 @@ package armdatafactory
 
 const (
 	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/datafactory/armdatafactory"
-	moduleVersion = "v7.0.0"
+	moduleVersion = "v8.0.0"
 )
 
 // ActivityOnInactiveMarkAs - Status result of the activity when the state is set to Inactive. This is an optional property
@@ -48,6 +48,22 @@ func PossibleActivityStateValues() []ActivityState {
 	}
 }
 
+// AmazonRdsForSQLAuthenticationType - The type used for authentication. Type: string.
+type AmazonRdsForSQLAuthenticationType string
+
+const (
+	AmazonRdsForSQLAuthenticationTypeSQL     AmazonRdsForSQLAuthenticationType = "SQL"
+	AmazonRdsForSQLAuthenticationTypeWindows AmazonRdsForSQLAuthenticationType = "Windows"
+)
+
+// PossibleAmazonRdsForSQLAuthenticationTypeValues returns the possible values for the AmazonRdsForSQLAuthenticationType const type.
+func PossibleAmazonRdsForSQLAuthenticationTypeValues() []AmazonRdsForSQLAuthenticationType {
+	return []AmazonRdsForSQLAuthenticationType{
+		AmazonRdsForSQLAuthenticationTypeSQL,
+		AmazonRdsForSQLAuthenticationTypeWindows,
+	}
+}
+
 // AzureFunctionActivityMethod - The list of HTTP methods supported by a AzureFunctionActivity.
 type AzureFunctionActivityMethod string
 
@@ -71,6 +87,66 @@ func PossibleAzureFunctionActivityMethodValues() []AzureFunctionActivityMethod {
 		AzureFunctionActivityMethodPOST,
 		AzureFunctionActivityMethodPUT,
 		AzureFunctionActivityMethodTRACE,
+	}
+}
+
+// AzureSQLDWAuthenticationType - The type used for authentication. Type: string.
+type AzureSQLDWAuthenticationType string
+
+const (
+	AzureSQLDWAuthenticationTypeSQL                           AzureSQLDWAuthenticationType = "SQL"
+	AzureSQLDWAuthenticationTypeServicePrincipal              AzureSQLDWAuthenticationType = "ServicePrincipal"
+	AzureSQLDWAuthenticationTypeSystemAssignedManagedIdentity AzureSQLDWAuthenticationType = "SystemAssignedManagedIdentity"
+	AzureSQLDWAuthenticationTypeUserAssignedManagedIdentity   AzureSQLDWAuthenticationType = "UserAssignedManagedIdentity"
+)
+
+// PossibleAzureSQLDWAuthenticationTypeValues returns the possible values for the AzureSQLDWAuthenticationType const type.
+func PossibleAzureSQLDWAuthenticationTypeValues() []AzureSQLDWAuthenticationType {
+	return []AzureSQLDWAuthenticationType{
+		AzureSQLDWAuthenticationTypeSQL,
+		AzureSQLDWAuthenticationTypeServicePrincipal,
+		AzureSQLDWAuthenticationTypeSystemAssignedManagedIdentity,
+		AzureSQLDWAuthenticationTypeUserAssignedManagedIdentity,
+	}
+}
+
+// AzureSQLDatabaseAuthenticationType - The type used for authentication. Type: string.
+type AzureSQLDatabaseAuthenticationType string
+
+const (
+	AzureSQLDatabaseAuthenticationTypeSQL                           AzureSQLDatabaseAuthenticationType = "SQL"
+	AzureSQLDatabaseAuthenticationTypeServicePrincipal              AzureSQLDatabaseAuthenticationType = "ServicePrincipal"
+	AzureSQLDatabaseAuthenticationTypeSystemAssignedManagedIdentity AzureSQLDatabaseAuthenticationType = "SystemAssignedManagedIdentity"
+	AzureSQLDatabaseAuthenticationTypeUserAssignedManagedIdentity   AzureSQLDatabaseAuthenticationType = "UserAssignedManagedIdentity"
+)
+
+// PossibleAzureSQLDatabaseAuthenticationTypeValues returns the possible values for the AzureSQLDatabaseAuthenticationType const type.
+func PossibleAzureSQLDatabaseAuthenticationTypeValues() []AzureSQLDatabaseAuthenticationType {
+	return []AzureSQLDatabaseAuthenticationType{
+		AzureSQLDatabaseAuthenticationTypeSQL,
+		AzureSQLDatabaseAuthenticationTypeServicePrincipal,
+		AzureSQLDatabaseAuthenticationTypeSystemAssignedManagedIdentity,
+		AzureSQLDatabaseAuthenticationTypeUserAssignedManagedIdentity,
+	}
+}
+
+// AzureSQLMIAuthenticationType - The type used for authentication. Type: string.
+type AzureSQLMIAuthenticationType string
+
+const (
+	AzureSQLMIAuthenticationTypeSQL                           AzureSQLMIAuthenticationType = "SQL"
+	AzureSQLMIAuthenticationTypeServicePrincipal              AzureSQLMIAuthenticationType = "ServicePrincipal"
+	AzureSQLMIAuthenticationTypeSystemAssignedManagedIdentity AzureSQLMIAuthenticationType = "SystemAssignedManagedIdentity"
+	AzureSQLMIAuthenticationTypeUserAssignedManagedIdentity   AzureSQLMIAuthenticationType = "UserAssignedManagedIdentity"
+)
+
+// PossibleAzureSQLMIAuthenticationTypeValues returns the possible values for the AzureSQLMIAuthenticationType const type.
+func PossibleAzureSQLMIAuthenticationTypeValues() []AzureSQLMIAuthenticationType {
+	return []AzureSQLMIAuthenticationType{
+		AzureSQLMIAuthenticationTypeSQL,
+		AzureSQLMIAuthenticationTypeServicePrincipal,
+		AzureSQLMIAuthenticationTypeSystemAssignedManagedIdentity,
+		AzureSQLMIAuthenticationTypeUserAssignedManagedIdentity,
 	}
 }
 
@@ -285,20 +361,6 @@ const (
 func PossibleDataFlowReferenceTypeValues() []DataFlowReferenceType {
 	return []DataFlowReferenceType{
 		DataFlowReferenceTypeDataFlowReference,
-	}
-}
-
-// DatasetReferenceType - Dataset reference type.
-type DatasetReferenceType string
-
-const (
-	DatasetReferenceTypeDatasetReference DatasetReferenceType = "DatasetReference"
-)
-
-// PossibleDatasetReferenceTypeValues returns the possible values for the DatasetReferenceType const type.
-func PossibleDatasetReferenceTypeValues() []DatasetReferenceType {
-	return []DatasetReferenceType{
-		DatasetReferenceTypeDatasetReference,
 	}
 }
 
@@ -811,20 +873,6 @@ func PossibleIntegrationRuntimeLicenseTypeValues() []IntegrationRuntimeLicenseTy
 	}
 }
 
-// IntegrationRuntimeReferenceType - Type of integration runtime.
-type IntegrationRuntimeReferenceType string
-
-const (
-	IntegrationRuntimeReferenceTypeIntegrationRuntimeReference IntegrationRuntimeReferenceType = "IntegrationRuntimeReference"
-)
-
-// PossibleIntegrationRuntimeReferenceTypeValues returns the possible values for the IntegrationRuntimeReferenceType const type.
-func PossibleIntegrationRuntimeReferenceTypeValues() []IntegrationRuntimeReferenceType {
-	return []IntegrationRuntimeReferenceType{
-		IntegrationRuntimeReferenceTypeIntegrationRuntimeReference,
-	}
-}
-
 // IntegrationRuntimeSsisCatalogPricingTier - The pricing tier for the catalog database. The valid values could be found in
 // https://azure.microsoft.com/en-us/pricing/details/sql-database/
 type IntegrationRuntimeSsisCatalogPricingTier string
@@ -909,20 +957,6 @@ func PossibleIntegrationRuntimeUpdateResultValues() []IntegrationRuntimeUpdateRe
 		IntegrationRuntimeUpdateResultFail,
 		IntegrationRuntimeUpdateResultNone,
 		IntegrationRuntimeUpdateResultSucceed,
-	}
-}
-
-// LinkedServiceReferenceType - Linked service reference type.
-type LinkedServiceReferenceType string
-
-const (
-	LinkedServiceReferenceTypeLinkedServiceReference LinkedServiceReferenceType = "LinkedServiceReference"
-)
-
-// PossibleLinkedServiceReferenceTypeValues returns the possible values for the LinkedServiceReferenceType const type.
-func PossibleLinkedServiceReferenceTypeValues() []LinkedServiceReferenceType {
-	return []LinkedServiceReferenceType{
-		LinkedServiceReferenceTypeLinkedServiceReference,
 	}
 }
 
@@ -1107,20 +1141,6 @@ func PossiblePhoenixAuthenticationTypeValues() []PhoenixAuthenticationType {
 		PhoenixAuthenticationTypeAnonymous,
 		PhoenixAuthenticationTypeUsernameAndPassword,
 		PhoenixAuthenticationTypeWindowsAzureHDInsightService,
-	}
-}
-
-// PipelineReferenceType - Pipeline reference type.
-type PipelineReferenceType string
-
-const (
-	PipelineReferenceTypePipelineReference PipelineReferenceType = "PipelineReference"
-)
-
-// PossiblePipelineReferenceTypeValues returns the possible values for the PipelineReferenceType const type.
-func PossiblePipelineReferenceTypeValues() []PipelineReferenceType {
-	return []PipelineReferenceType{
-		PipelineReferenceTypePipelineReference,
 	}
 }
 
@@ -1344,6 +1364,22 @@ func PossibleSQLAlwaysEncryptedAkvAuthTypeValues() []SQLAlwaysEncryptedAkvAuthTy
 	}
 }
 
+// SQLServerAuthenticationType - The type used for authentication. Type: string.
+type SQLServerAuthenticationType string
+
+const (
+	SQLServerAuthenticationTypeSQL     SQLServerAuthenticationType = "SQL"
+	SQLServerAuthenticationTypeWindows SQLServerAuthenticationType = "Windows"
+)
+
+// PossibleSQLServerAuthenticationTypeValues returns the possible values for the SQLServerAuthenticationType const type.
+func PossibleSQLServerAuthenticationTypeValues() []SQLServerAuthenticationType {
+	return []SQLServerAuthenticationType{
+		SQLServerAuthenticationTypeSQL,
+		SQLServerAuthenticationTypeWindows,
+	}
+}
+
 // SalesforceSinkWriteBehavior - The write behavior for the operation. Default is Insert.
 type SalesforceSinkWriteBehavior string
 
@@ -1475,22 +1511,6 @@ func PossibleScriptActivityParameterTypeValues() []ScriptActivityParameterType {
 		ScriptActivityParameterTypeSingle,
 		ScriptActivityParameterTypeString,
 		ScriptActivityParameterTypeTimespan,
-	}
-}
-
-// ScriptType - The type of the query. Type: string.
-type ScriptType string
-
-const (
-	ScriptTypeNonQuery ScriptType = "NonQuery"
-	ScriptTypeQuery    ScriptType = "Query"
-)
-
-// PossibleScriptTypeValues returns the possible values for the ScriptType const type.
-func PossibleScriptTypeValues() []ScriptType {
-	return []ScriptType{
-		ScriptTypeNonQuery,
-		ScriptTypeQuery,
 	}
 }